Simplify flag lex+gramm, sync flags, fix posting
These changes use a default rule instead of user action to switch back
from the possible flag start condition to INITIAL.  The possible flag
condition has been changed to
also occur after an indent, to
recognize posting flags, and renamed
accordingly to POSSIBLEFLAG.

Asterisk, colon, and hash lexing has been simplified: they are now
directly recognized as flags in the lexer when they are flags, removing
the need for checks in the grammar.

The lexer has been changed to only allow flags that are in
beancount.core.flags, as described in the comments and language syntax.
It has also been changed to no longer allow
arbitrary flags at the beginning of the line
as a way to skip that line.
(https://github.com/beancount/beancount/issues/516)

Note that these last two changes might break files that currently work,
but don't actually comply with the language syntax documentation.